- Migrate from JPA to R2DBC in tests and main application dependencies - Replace `MockMvc` with `WebTestClient` in end-to-end test cases - Refactor tests to support reactive repositories and coroutine-based setup

+++ src/main/kotlin/org/my/firstcircletest/delivery/http/dto/request/CreateUserRequestDto.kt
@@ -13,7 +13,7 @@ data class CreateUserRequestDto(
     @param:JsonProperty("init_balance")
     @field:Min(value = 0, message = "Initial balance cannot be negative")
     @field:Max(100_000_000, message = "Requested initial balance exceeds the maximum limit")
-    val initBalance: Int = 0
+    val initBalance: Long = 0
 ) {
     fun toDomain() = CreateUserRequest(
         name = this.name,
