+++ src/main/kotlin/org/my/firstcircletest/domain/entities/CreateUserRequest.kt
@@ -2,5 +2,5 @@ package org.my.firstcircletest.domain.entities
 
 data class CreateUserRequest(
     val name: String,
-    val initBalance: Int
+    val initBalance: Long
 )
