+++ src/main/kotlin/org/my/firstcircletest/domain/entities/CreateWalletRequest.kt
@@ -2,5 +2,5 @@ package org.my.firstcircletest.domain.entities
 
 data class CreateWalletRequest(
     val userId: UserID,
-    val balance: Int
+    val balance: Long
 )
